move `disabling menus from cms` tests to `test_backend` and removed compatibility with older python versions  than 3.4 for this specific test
